remove manual URL construction from <GitCommitNode>, allow showing repository
Previously, the `<GitCommitNode>` component accepted a `repoName` prop that it used to manually construct a URL. This is a bad practice; it should just use the URL from the `GitCommit` GraphQL value it receives.

Also, adds an option for this component to show the repository the commit is from (which is useful when showing commits from multiple repositories together).