simplify frontend CSS build in Webpack (#5266)
Previously, our Webpack build had a separate set of entrypoints for CSS (vs. TypeScript). It also had a hack to avoid a red-herring 404 in the devtools browser console when running Sourcegraph in local dev (which does not produce `.css` output files; it uses `style-loader` to inject styles via JavaScript so we get hot-reloading, ie instant feedback for CSS changes).

Now, it only has a single set of (TypeScript) entrypoints that, in turn, import the relevant CSS files. Also, the hack was removed in favor of not emitting a `<link rel="stylesheet" ...>` tag when running Sourcegraph in local dev.