misc GraphQL API code cleanup (#5279)
* add doc comment file to graphqlbackend package (enterprise)

* in GraphQL access tokens connection, only return the requested number of nodes

* allow other packages contributing GraphQL resolvers to use ExternalLinks

If a package wants to create its own ExternalLink GraphQL value, it needs to be able to instantiate `externallink.Resolver`. That type is a struct with only unexported fields, so this `NewResolver` func is necessary.

* move registry side-effects import to shared

This package *is* indeed imported and required by both the OSS and enterprise builds, so it should be in shared. In practice, this does not make any difference in the final binary because the enterprise build already imported this package. But moving it to the correct place (shared) is still good because it documents the intent (for it to be shared, not for it to be OSS-only).