authz: Do not filter repo ids to cache in Bitbucket Server provider
Up until now, the cached set of repository ids that a user is allowed to
see was dependent on kind of request first done when there was not
cache (or it had expired). This commit changes the code so that we store
all authorized repository ids in the cache, not only those that were to
be verified when RepoPerms was called.