repo-updater: update aws codecommit module
AWS now just takes a Doer interface for the HTTP client. We now can't get at the
http.Client used by AWS. However, we can get the transport. We then configure
the transport in the same way. What is missing is the customizations to the
client (has to do with retries).