web: Render external service form errors/warnings as Markdown
This is a really simple first fix for #4842, as suggested by
@felixbecker.

The _proper_ solution would probably be to make the errors themselves
more readable & easier to parse, but that's a lot more work than this
fix here.

Until we find the time/muse to do that, this here already improves on
the current sate.