prevent double-highlighting
This fixed a bug I saw where rerenders of an ancestor component caused highlights to be applied multiple times. This resulted in the background color of the highlighted text getting darker and darker because it had multiple wrapping elements with partial opacity.

The `highlightNodeHelper` has a check to prevent this case, but it was not effective.