expose common namespaceName in GraphQL API for Namespace (User, Org) (#5395)
Both User and Org in our GraphQL API implement Namespace, which is container for certain types of data and settings (currently only saved searches, but intended to be more for a8n).

The User and Org namespace is shared; i.e., a user and org can't have the same name. (This is enforced in the DB by using the `names` table.) API clients retrieving data from multiple namespaces (e.g., a list of all saved searches, or a8n campaigns in the future) can thus use the `namespaceName` as a unique name prefix for the data they retrieve.

Exposing the `namespaceName` as the same-named field on both User and Org makes this clearer. API clients could type-switch and use `User.username` or `Org.name`, but this is clearer.