Allow user-select in ResultContainer header
Often my goal in doing a search in Sourcegraph is to find files where I need to make some kind of changes in my editor. To that end, my lack of ability to copy filenames from headers is super frustrating, so much so that I came here to submit this feature request. I know I'm not the only one of my coworkers who has been frustrated by this.

Please let me know if there are further things to do to help move this along.