hide quicklinks bar if empty
Fixes an issue where an empty horizontal bar would be shown on the search results page even if there were no quicklinks. Also simplifies the quicklinks component usage a bit so that callers need not check whether quicklinks are empty (this check was omitted on the search results page, which caused the bug).