always use root-relative names in Git entries
os.FileInfos representing Git objects now always return the full path relative to the repository root in the Name method. Previously the behavior depended on several complex factors, such as whether the entry was the root, whether the ReadDir traversal was recursive, etc.

This also lets us remove a field in gitTreeEntryResolver.