cache repo language inventory per-tree to improve hit rate
Previously, we cached the result of a repository inventory call (which detects languages of the repository's files and sorts them by total number of bytes of each language) by repo+commit. This meant that each commit in a repository required recomputing the full inventory.

When computing the inventory was nearly instant, this was acceptable. However, we introduced an enhanced repository inventory (currently behind the USE_ENHANCED_LANGUAGE_DETECTION feature flag, off by default) that is significantly slower (and more accurate) because it uses file contents to detect the language of a file.

Because of this, we needed to improve the perf of the repo inventory. Caching based on Git tree OIDs means that we (usually) can reuse most of the work between 2 commits in the same repo, under the assumption that most Git trees are unchanged between commits.

The following benchmark results show the perf characteristics. It takes about the same amount of time to compute the repo inventory for 1 commit of kubernetes, 2 commits of kubernetes, and 100 commits of kubernetes. Each extra commit to compute (on the margin) adds very little extra time.

```
$ REPO=github.com/kubernetes/kubernetes N=1 bash ~/tmp/benchmark-inventory.bash

real	0m17.099s

$ REPO=github.com/kubernetes/kubernetes N=2 bash ~/tmp/benchmark-inventory.bash

real	0m18.318s

$ REPO=github.com/kubernetes/kubernetes N=100 bash ~/tmp/benchmark-inventory.bash

real	0m45.838s

```