e2e-tests: Reduce flakiness of Java symbol bar test
This addresses the recent flakiness of this test by by choosing a
smaller Java repository and file with less symbols that all fit into the
symbol bar.

The assumption underlying this change is explained in this Slack
message:

    https://sourcegraph.slack.com/archives/C07KZF47K/p1567537846051500?thread_ts=1567411847.037700&cid=C07KZF47K

Before this change, the number of symbols in the given subfolder was so
high that only a subset could be displayed (to see the others, one would
have to scroll).

The suspicion is that this list of symbols is unordered, which meant
that sometimes the desired symbol was visible and sometimes not.

This change here switches from the `gitblit/gitblit` repo to
`sourcegraph/java-langserver` and changes the subfolder to be one that
contains only the 19 symbols found in the single file. In the configured
viewport, all of these symbols are visible, independent of sort order.