e2e-tests: Make symbol bar tests for Go more stable
After multiple flaky e2e failures on `master`, this does for the Go
symbol bar tests what #5474 did for the Java symbol bar tests.

It changes the test "fixtures" to use directories/files with less
symbols so that the content of the symbol bar is stable, even if the
ordering of the symbols is not.