search: Introduce Indices for sharding indexed search
Indices will contain functionality related to splitting up searching and
indexing across multiple shards. This commit introduces the use of our
endpoint pkg to shard repositories amongst indexed search pods.

The hostname an indexed search pod advertises to us will likely be different
to the address we use to contact it. As such we have logic to match it up with
an endpoint address.

The endpoint package also allows us to use service discovery in the future (like
we do with searcher and k8s). However, we won't use it yes since it will require
more extensive testing first.