a8n: DB trigger to clean up references to campaigns/changesets on deletion
Before this change, if you'd remove an `external_service` that delete
would cascade and also delete the `repo`s associated with the
`external_service`.

That in turn would also cascade and delete the `changeset` whose
`repo_id` matches the deleted `repo`.

There are still a lot of unanswered questions around deletion in the a8n
subparts of the system (e.g.: should we delete a campaign if the
external service was deleted, and possibly recreated?).

Until we answer those and come up with a full strategy around deletion,
we try to at least not leave invalid state around when we delete
something.

To do that without switching to a JOIN table we add two triggers to the
`changesets` and `campaigns` tables respectively that delete references
to each other in case a row gets deleted.

Example: if a `changeset` is deleted, the trigger will remove its ID
from the `changeset_ids` columns of `campaigns` to which it has been
added.