e2e-tests: Add external service test for Bitbucket Server
This automates entry 9.4 from testing plan for 3.8. And it also
automates half of 9.12, by not using a `repositoryQuery` in the config.

Just like the test for AWS CodeCommit, it only runs if the environment
variables are set. I will do this for our buildkite configuration.
Locally the test already passes with the BBS config out of
`dev-private`.