e2e-tests: Test that `repositoryQuery` is not needed if `repos` is set
This automates the second half of 9.12: by removing the
`repositoryQuery` here we test that the config is valid with `repos`
only. In other tests we test that `repositoryQuery` without `repos` is
valid.

The existing BitBucket Server test from f88f6a9f9b tests that this holds
true for BBS as well.