bitbucketserver: Ignore host when replaying API fixtures
On CI we now set BITBUCKET_SERVER_HOST, which has lead to CI failing since we
now attempt to speak to the host. We speak to the host since the URLs now don't
match what is recorded in our fixtures. This commit updates the matching logic
so that we ignore the host and scheme, and instead only rely on the Path for
matching a request to a fixture.