repo-updater: Syncer.Diffs -> Syncer.Synced
The only reader of the diffs channel would never actually look at the
individual fields of diff, but would straight away convert it to a Repos. We
instead update the channel to be a Repos.

This will make it possible to stream out changes where we can't create as
accurate diffs.