e2e tests: Automate 9.4 and 9.12 from the release testing grid (#5629)
* e2e-tests: Add external service test for Bitbucket Server

This automates entry 9.4 from testing plan for 3.8. And it also
automates half of 9.12, by not using a `repositoryQuery` in the config.

Just like the test for AWS CodeCommit, it only runs if the environment
variables are set. I will do this for our buildkite configuration.
Locally the test already passes with the BBS config out of
`dev-private`.

* e2e-tests: Test that `repositoryQuery` is not needed if `repos` is set

This automates the second half of 9.12: by removing the
`repositoryQuery` here we test that the config is valid with `repos`
only. In other tests we test that `repositoryQuery` without `repos` is
valid.

The existing BitBucket Server test from f88f6a9f9b tests that this holds
true for BBS as well.

* e2e-tests: Retry assertions for new external service repos

* e2e-tests: Use ensureRepos in external service tests