repo-updater: Refactor Syncer to be created without constructor (#5646)
This is a pure refactoring (no functionality change) to make it possible to create Syncer without a constructor. This is useful since in many test cases we set nil for values we don't care about, leading to quite an ugly constructor call. This will also facilitate introduction of more fields.