repo-updater: Stream inserts (#5645)
Syncing requires fetching from all sources before it can update the database.
For large instances this leads to a long time (minutes) from the time an admin
makes a change to an external service to seeing the effects. The most common
complaint is adding repositories to the external service, and not seeing
them. This commit will eagerly add new repositories into the store before
running the full sync. This is a much smaller change than making the sync
engine fully stream based.