repo-updater: Stream GitHub last configs in lists first
This commit is making the observation that admins usually add
repos/configurations to the end of lists. We now eagerly insert new
repositories. So by processing GitHub configurations such that we first stream
the end of the repos list, will likely lead to the new repository appearing
sooner. The same assumption is applied to repository queries and orgs.