searcher: don't copy regular expressions (#5642)
It is no longer necessary to copy regular expressions for concurrent access as of go 1.12

SA1019: rg.re.Copy is deprecated: In earlier releases, when using a Regexp in multiple goroutines, giving each goroutine its own copy helped to avoid lock contention. As of Go 1.12, using Copy is no longer necessary to avoid lock contention. Copy may still be appropriate if the reason for its use is to make two copies with different Longest settings.  (staticcheck)