e2e: wait for signup form before filling fields
It appears that [this build](https://buildkite.com/sourcegraph/sourcegraph/builds/43598#a0be2ad6-0cab-4ff2-a443-2ea58b1edb09) failed because the email field didn't exist on the page when puppeteer attempted to fill it on `/site-admin/init`.

This ensures the signin/signup forms are present on the page before attempting to fill any fields.