a8n: Setup rcache for tests
We accidently relied on this being setup by another test when these
tests were still in the `graphqlbackend` package. Now that they've
moved, the `rcache` used by the `GitHubSource` was not setup for
testing.

The result was that the tests took ~16 seconds to run on CI until the
rcache connection times out (without an error, since it's a cache) and
90+ seconds on my machine.