a8n: Fix reading Bitbucket Server changesets from DB
This is a follow-up to https://github.com/sourcegraph/sourcegraph/pull/5768

I forgot to extend this switch statement. And since we previously didn't
read back the created changeset from the database with the metadata
fields, it didn't cause a problem.

In this change I extended the test suite so that it fails when we read
an "unknown" changeset from the DB and extended the switch statement.