graphqlbackend: Only sort branches if less than 1000 (#5831)
To sort branches we need to fetch each branch's commit individually. This is a
slow operation. So we introduce a new field to the GraphQL calls
"interactive". It defaults to true, and will skip on the expensive operation
if we have too many commits to get. 1000 is a thumb-sucked number.