lsif: use proper sourcegraph/alpine base image
Prior to this change, lsif-server was not using our `sourcegraph/alpine` base
image. Using this base image and not standard `alpine` is very important for
three reasons:

1. It installs [a critical package](https://github.com/sourcegraph/sourcegraph/blob/10e81ba9fa9f055cae8e755682dc74ef4413204c/docker-images/alpine/Dockerfile#L17-L25) without which DNS lookups fail in certain deployment contexts.
2. It [creates a standard `sourcegraph` user](https://github.com/sourcegraph/sourcegraph/blob/10e81ba9fa9f055cae8e755682dc74ef4413204c/docker-images/alpine/Dockerfile#L10-L15) with a static UID/GID, which is important for `deploy-sourcegraph-docker` deployment contexts where file permissions must sometimes be set manually on the host.
3. It unifies the base image of alpine and packages we install across all services, e.g. `tini`.

Additionally, because lsif-server had no user before, it is being ran by `root`
currently and all of its files are owned by `root` in Kubernetes deployments. I
will send a pair change with this one to automatically update the permissions
in such clusters.