internalapi: Allow ping to wait on other services (#5866)
We want zoekt-indexserver to only start once gitserver is up. It doesn't know
how to communicate with gitserver directly, so we need to add a way for it to
detect gitserver being available. It already pings the frontend, it can now
additionally specify gitserver as a ping.

Alternatively I considered adding a new internal endpoint, but this seemed
like overkill and ping already fits this purpose.

Note: indexserver already correctly handles gitserver not being up, but it
leads to a lot of logspam. EG when you start up server you will often see a
lot of "resolve-rev 500" logs from the frontend. This is because indexserver
is operating before gitserver is ready.