ci: Set coverpkg to record dependency code coverage
We will now record coverage in all sourcegraph packages. Previously we only
recorded the coverage per package test, so missed things like our DB tests
covering the migrations bindata.

We used to set this flag, but I think we lost it during our OSS
process. Additionally it had a bug that was introduced and then fixed again in
go1.13.