repo-updater: Deterministic pagination from github and gitlab API (#5874)
Previously we did this so we would add recently pushed repositories first.
However, with paged based pagination if a repository is pushed to while we are
querying we may miss it. This is the likely cause of repositories occasionally
being deleted at a customer.

Instead we order by created. In the case of GitLab this is the default so we don't specify it.
If a repo is removed from upstream while syncing we could delete a repository.
However, this is much less likely to occur so is a good enough solution for now.

Updated testdata with the following commands:

  ruplacer --go sort=pushed\& 'sort=created&'
  ruplacer --go order_by=last_activity_at\& ''