Browser extension GSuite deployment fixes (#5894)
* Remove unused storage items

- Remove unused repoLocations storage items.
- Remove `enterpriseUrls`: the only consequence of having enterpriseURLs present in storage was to set the browser action to open the options page instead of the popup. In the options page, nothing specific is done when enterpriseURLs were present. They were otherwise unused in the browser extension.

* Handle reading the sourcegraphURL from managed storage

Adds a new property to the managed storage URL, sourcegraphURL. Changes to the storage.managed.sourcegraphURL are observed throughout the lifetime of the extension, and copied to sync storage.

* Storage: split StorageItems by storage area

There are now different storage item types for local, sync and managed storage items. The typings of `observeStorageKey` and `storage` reflect this. Rather than copying all items from managed storage to sync storage upon installation, relevant storage items are directly observed from managed storage (for instance, `phabricatorMappings`).

* Use observeSourcegraphURL everywhere

Changes all calls to `observeStorageKey('sync', 'observeSourcegraphURL')` to `observeSourcegraphURL()` so that the sourcegraphURL is correctly defaulted to https://sourcegraph.com if not set in storage.