gitserver: Reclone repositories with gic gc warnings
45 days is too long to wait for large repositories such as kubernetes. Git
will add a warning log file called `gc.log` when it detects a clone is
needed. Our reclone job now takes this signal into account when deciding to
reclone. If the file is present and it has been atleast a day since our last
clone, the repository will get recloned.