a8n: Fix listing all ChangesetEvents resulting in empty results
This fixes #6011.

The panic in the bug report was caused by `len(cs)` being `0` and
trying to access `cs[:-1]`, leading to an index-out-of-range panic.

This happens only when `Limit` has been set to `-1` before (it's
incremented to `0` in the function that constructs the query) to list
all changeset events and the query resulting in an empty result set.