warn users if they may be accidently searching literally for quotes
This is an issue that is hindering my personal workflow and happens to me
multiple times a day even though I "know" it is the new behavior it still
trips me up constantly. I fear our users will experience the same pain, and
a fix was simple enough so I went ahead and implemented this.

If you have performed a search literally and it contains quotes, we now
display a note that the quotes are also being interpreted literally to
help users like myself who are used to "quoting" everything for literal
search:

![image](https://user-images.githubusercontent.com/3173176/67041871-77355a00-f0db-11e9-8115-6c5f69602cf7.png)

Fixes #6016