ci: Only E2E run for releases
PRs and master will no longer run E2E tests. This should improve reliability
and speed of CI for the general case, while still ensuring e2e tests pass on
tagged releases.

The downside is we won't notice severe regressions to e2e, so will have to
deal with the issues once we do a release. But as it stands the e2e tests are
in a bad shape and no one is fixing them. We are building a pavlovian response
to e2e failing and the retry button.