Provide option to override Sourcegraph host value for src-expose
Currently, the config output when running src-expose has the host in the external configuration hard-coded to `host.docker.internal`.

This change, allows for the overriding of this value, so if the sourcegraph host is known, e.g. sourcegraph.internal.example.com, this can be provided so the external service configfuration does not have to be modified.