graphqlbackend: Introduce JSONCString scalar type (#6209)
This is the first step towards implementing #6135.

In order to keep the pull requests and diffs as small was possible, I
split this out, since it's generic and can already be merged.

I manually tested editing the site configuration, critical config in the
management console, user and org settings. External service
configuration is tested by e2e tests.