lsif: Fix `no such file or directory [...] search-jobs.lua`
When I run `./enterprise/dev/start.sh` I got the following output:

    lsif-server | ERROR failed to start process { service: 'lsif-server',
    lsif-server |   error:
    lsif-server |    { Error: ENOENT: no such file or directory, open '/Users/thorstenball/work/sourcegraph/lsif/out/search-jobs.lua'
    lsif-server |      errno: -2,
    lsif-server |      code: 'ENOENT',
    lsif-server |      syscall: 'open',
    lsif-server |      path:
    lsif-server |       '/Users/thorstenball/work/sourcegraph/lsif/out/search-jobs.lua' } }

I assumed it was because this

https://github.com/sourcegraph/sourcegraph/blob/61c29221e06580e831d6119ab16cfe524928c4f4/lsif/src/server.ts#L597

is relative to the location of the `server.ts` file and when that is
compiled to `out/server.js` the path is wrong.

I don't know enough about the TypeScript build process to be sure that
my change here is correct and won't break with a different build config,
but, well, I thought I'd give it a shot.