frontend: Introduce serveIndex
This is a specialised handler for zoekt indexserver. Previously we used a
generic "reposList" endpoint which ended up being specialized for zoekt. We
now make this clear by calling the endpoint Index. Currently it is a
copy-paste of serveRepoList, with some changes. It has an API with a targetted
responses (a list of repo names vs list of repo structs).

Additionally this endpoint receives a list of repositories currently indexed
by Zoekt. The next commit will use that list.