search: Zoekt sends list of indexed repos when requesting what to index (#6290)
Previously there was a race condition when computing the list of repositories
to index. We try to avoid dropping repos that are already indexed by a node if
no other node has indexed it. We did this by asking zoekt-webserver the list
of repos. However, zoekt-indexserver may ask for the list of repos before
zoekt-webserver has started => this list being empty.