frontend: Return errors for internal REST API calls
I was trying to debug an issue via curling the internal apis in production.
However, I would only get back a `500` with no information. To find the error
message I had to look in logs. Rather it makes sense to return it. This will
also allow users of the internal API to propogate the reason for failure.