server: Automatically repair redis AOF (#6230)
If the AOF file exists, we automatically run the repair command on it. We do
it in a best-effort fashion. So if the repair fails we will still try start
redis, but will include in the log the failure. This should help a common
reason our server container won't start. For example this has happened to
sourcegraph.sgdev.org twice in the last week.