endpoint: Ignore empty Hostname from endpoints
In production we are getting empty hostnames from the searcher endpoints. When
manually running `kubectl get endpoints -o json searcher` the `hostname` field
is not set. However, the k8s client library we are using may be marshalling
the field anyways for some reason. Either way this is causing searcher
unavailability in production.