move non-technical docs to about repository
As part of our goal to become handbook-first, this deletes Markdown documents that are not related to the code or product. These are being moved to the https://github.com/sourcegraph/about repository. Keeping all company, handbook, and process documentation in that repository makes it easier for everyone to contribute to and allows us to render information more nicely on about.sourcegraph.com (e.g., to have our product roadmap render as something other than just Markdown-rendered HTML).

Also sets up redirects for deleted pages.