internal/vcs/git: Introduce blobReader
Added a blobReader struct which supports the io.ReadCloser interface.
This allows us to more easily expose an API allowing streaming reads of
blob contents.

ReadFile was modified to use the blobReader.

Error wrapping that was previously done in readFileBytes was moved into
blobReader.

Instead of returning a nil error when reading from a submodule we now
return io.EOF.