a8n: Filter out repositories of unsupported external services
We currently only support GitHub and Bitbucket Server.

It doesn't make sense and only leads to bugs if we try to run
CampaignJobs over repositories whose external service we don't support,
because we later can't turn those CampaignJobs into Changesets on the
codehost.

And before manually creating a changeset we check that all repositories
are supported.